Move classify api under /api/v1 router
