+++ pebblo/app/api/api.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter, Depends
 
 from pebblo.app.config.config import var_server_config_dict
-from pebblo.app.service.classification import Classification
 from pebblo.app.service.prompt_gov import PromptGov
 from pebblo.app.utils.handler_mapper import get_handler
 
@@ -48,10 +47,4 @@ class App:
         # "/prompt/governance" API entrypoint
         prompt_obj = PromptGov(data=data)
         response = prompt_obj.process_request()
-        return response
-
-    @staticmethod
-    def classify_data(data: dict):
-        cls_obj = Classification(data)
-        response = cls_obj.process_request()
         return response
