+++ pebblo/app/routers/routers.py
@@ -3,9 +3,11 @@ Routes for Pebblo Server
 """
 
 from pebblo.app.api.api import App
+from pebblo.app.api.v1 import APIv1
 
 # Create an instance of APp with a specific prefix
 router_instance = App(prefix="/v1")
+api_v1_router_instance = APIv1(prefix="/api/v1")
 
 # Add routes to the class-based router
 router_instance.router.add_api_route(
@@ -33,9 +35,9 @@ router_instance.router.add_api_route(
     response_model_exclude_none=True,
 )
 
-router_instance.router.add_api_route(
+api_v1_router_instance.router.add_api_route(
     "/classify",
-    App.classify_data,
+    APIv1.classify_data,
     methods=["POST"],
     response_model=dict,
     response_model_exclude_none=True,
